requirements.txt
crewai
langchain-google-genai
fastapi
uvicorn
python-dotenv
google-generativeai
langgraph
langchain-core
